face_recognition_app/flask_api/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
hypothesis>=6.0.0
deepface>=0.0.79
mtcnn>=0.1.1
tensorflow>=2.13.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
Pillow>=10.0.0
tf-keras
retinaface